perbaikan bug nominal simpanan

+++ database/migrations/2026_05_01_133525_create_simpanan_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
     $table->foreignId('anggota_id')->constrained('anggota')->onDelete('cascade');
 
     $table->enum('jenis_simpanan', ['wajib', 'pokok', 'sukarela']);
-    $table->integer('jumlah');
+    $table->decimal('jumlah', 30, 0);
     $table->date('tanggal');
 
     $table->timestamps();
